+++ DataKitsServer/src/main/java/io/github/xuyao5/datakitsserver/job/File2EsDemoJob.java
@@ -4,9 +4,9 @@ import io.github.xuyao5.datakitsserver.configuration.elasticsearch.EsKitsConfig;
 import io.github.xuyao5.datakitsserver.vo.MyFileDocument;
 import io.github.xuyao5.dkl.eskits.service.File2EsService;
 import io.github.xuyao5.dkl.eskits.service.config.File2EsConfig;
+import io.github.xuyao5.dkl.eskits.support.auxiliary.AliasesSupporter;
+import io.github.xuyao5.dkl.eskits.support.auxiliary.CleaningSupporter;
 import io.github.xuyao5.dkl.eskits.support.batch.ReindexSupporter;
-import io.github.xuyao5.dkl.eskits.support.boost.AliasesSupporter;
-import io.github.xuyao5.dkl.eskits.support.boost.CleaningSupporter;
 import io.github.xuyao5.dkl.eskits.support.general.IndexSupporter;
 import io.github.xuyao5.dkl.eskits.util.DateUtilsPlus;
 import io.github.xuyao5.dkl.eskits.util.FileUtilsPlus;
